+++ IT2B-G05-BankApp/it2bank/src/main/java/nl/inholland/it2bank/model/dto/BankAccountDTO.java
@@ -2,7 +2,7 @@ package nl.inholland.it2bank.model.dto;
 
 import nl.inholland.it2bank.model.BankAccountModel;
 
-public record BankAccountDTO(String iban, Integer ownerId, Integer statusId, Double balance, Integer absoluteLimit, Integer typeId) {
+public record BankAccountDTO(String iban, Long ownerId, Integer statusId, Double balance, Integer absoluteLimit, Integer typeId) {
     public BankAccountDTO(BankAccountModel bankAccountModel) {
         this(bankAccountModel.getIban(),
                 bankAccountModel.getOwnerId(),
